+++ app/src/main/java/com/example/ejt7/dataBase/dao/DAO.kt
@@ -6,5 +6,5 @@ interface DAO <T>{
     fun findAll(context: Context?): List<T>
     fun save(context: Context?, t:T)
     fun update(context: Context?, t:T)
-    fun delete(context: Context?, id:Int)
+    fun delete(context: Context?, id:Long)
 }
